Fixed that pile of errors by actually clean up the project.
