+++ DebugConsole/DEBUG_TIMEBASED/DebugConsole.tlog/DebugConsole.lastbuildstate
@@ -1,2 +1,2 @@
 #TargetFrameworkVersion=v4.0:PlatformToolSet=v142:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0
-DEBUG_TIMEBASED|Win32|C:\Users\Pegasus\source\repos\Tetris_ML\Core\|
+DEBUG_TIMEBASED|Win32|C:\Users\Pegasus\source\repos\Pegasust\Tetris_ML\Core\|
