+++ DebugConsole/keyboard_input.cpp
@@ -1,6 +1,6 @@
 #include "keyboard_input.h"
 
-void keyboard_input::get_input()
-{
-	int input = getch();
-}
+//void keyboard_input::get_input()
+//{
+//	int input = getch();
+//}
